refactor away the collector base
