+++ lib/coverband/reporters/simple_cov_report.rb
@@ -43,7 +43,7 @@ module Coverband
           access_key_id: Coverband.configuration.s3_access_key_id,
           secret_access_key: Coverband.configuration.s3_secret_access_key
         }
-        Coverband::Adapters::S3ReportWriter.new(Coverband.configuration.s3_bucket,
+        Coverband::Utils::S3ReportWriter.new(Coverband.configuration.s3_bucket,
                                                 s3_writer_options).persist! if Coverband.configuration.s3_bucket
       end
     end
